Move test mock to separate testing file
